Created new suspend.py file that contains the suspend mode feature, removed it from tools.py and minor changes in app.py file.
